refactor: splitted the footer contents component
